# src/AppForSEII2526.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Purchase/GetPurchase/{id}": {
      "get": {
        "tags": [
          "Purchase"
        ],
        "operationId": "GetPurchase",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Purchase/CreatePurchase": {
      "post": {
        "tags": [
          "Purchase"
        ],
        "operationId": "CreatePurchase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PaymentMethod": {
        "enum": [
          "CreditCard",
          "PayPal"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "PurchaseDetailDTO": {
        "required": [
          "customerName",
          "customerSurname",
          "deliveryAddress",
          "paymentMethod"
        ],
        "type": "object",
        "properties": {
          "deliveryAddress": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string"
          },
          "purchaseItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseItemDTO"
            },
            "nullable": true
          },
          "totalPrice": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "customerName": {
            "maxLength": 25,
            "minLength": 2,
            "type": "string"
          },
          "customerSurname": {
            "maxLength": 25,
            "minLength": 2,
            "type": "string"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "purchaseDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PurchaseForCreateDTO": {
        "required": [
          "customerName",
          "customerSurname",
          "deliveryAddress",
          "paymentMethod"
        ],
        "type": "object",
        "properties": {
          "deliveryAddress": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string"
          },
          "purchaseItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseItemDTO"
            },
            "nullable": true
          },
          "totalPrice": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "customerName": {
            "maxLength": 25,
            "minLength": 2,
            "type": "string"
          },
          "customerSurname": {
            "maxLength": 25,
            "minLength": 2,
            "type": "string"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          }
        },
        "additionalProperties": false
      },
      "PurchaseItemDTO": {
        "required": [
          "brand",
          "color",
          "model",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "deviceID": {
            "type": "integer",
            "format": "int32"
          },
          "brand": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "model": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "color": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string"
          },
          "priceForPurchase": {
            "type": "number",
            "format": "double"
          },
          "quantity": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}